// admin-dashboard/src/app/core/api-client/model/onboarding-status-response.ts
/**
 * CoreCourierService.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OnboardingStatusResponse { 
    tenantId?: string | null;
    profileCompleted?: boolean;
    ratesCompleted?: boolean;
    status?: string | null;
}